# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quickstart in Couchbase with PHP and Laravel",
...
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/airports/list": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get list of airports",
                "description": "Returns list of airports",
                "operationId": "getAirportsList",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Airport"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airports/{id}": {
            "get": {
                "tags": [
                    "Airports"
                ],
                "summary": "Get airport information",
                "description": "Returns airport data",
                "operationId": "getAirportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Airport"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Airports"
                ],
                "summary": "Update an existing airport or create a new one if it does not exist",
                "description": "Update an existing airport or create a new one if it does not exist",
                "operationId": "updateAirport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Airport object that needs to be updated or created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Airport"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Resource created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Create a new airport",
                "description": "Create a new airport",
                "operationId": "createAirport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Airport object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Airport"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Airports"
                ],
                "summary": "Delete an airport",
                "description": "Delete an airport",
                "operationId": "deleteAirport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Airport": {
                "title": "Airport",
                "required": [
                    "airportname",
                    "city",
                    "country",
                    "icao",
                    "tz",
                    "geo"
                ],
                "properties": {
                    "airportname": {
                        "description": "Name of the airport",
                        "type": "string"
                    },
                    "city": {
                        "description": "City where the airport is located",
                        "type": "string"
                    },
                    "country": {
                        "description": "Country where the airport is located",
                        "type": "string"
                    },
                    "faa": {
                        "description": "FAA code of the airport",
                        "type": "string"
                    },
                    "icao": {
                        "description": "ICAO code of the airport",
                        "type": "string"
                    },
                    "tz": {
                        "description": "Time zone of the airport",
                        "type": "string"
                    },
                    "geo": {
                        "description": "Geographical coordinates of the airport",
                        "properties": {
                            "lat": {
                                "description": "Latitude",
                                "type": "number",
                                "format": "float"
                            },
                            "lon": {
                                "description": "Longitude",
                                "type": "number",
                                "format": "float"
                            },
                            "alt": {
                                "description": "Altitude",
                                "type": "number",
                                "format": "float"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    }
}